+++ FBSDKLoginKit/FBSDKLoginKit/FBSDKLoginConfiguration.h
@@ -44,8 +44,27 @@ NS_SWIFT_NAME(LoginConfiguration)
 /// The requested permissions for the login attempt. Defaults to an empty set.
 @property (nonatomic, readonly, copy) NSSet<FBSDKPermission *> *requestedPermissions;
 
+/// The Messenger Page Id associated with this login request.
+@property (nonatomic, readonly, copy, nullable) NSString *messengerPageId;
+
 - (instancetype)init NS_UNAVAILABLE;
 + (instancetype)new NS_UNAVAILABLE;
+
+/**
+ Attempts to initialize a new configuration with the expected parameters.
+
+ @param permissions the requested permissions for a login attempt. Permissions must be an array of strings that do not contain whitespace.
+ The only permissions allowed when the `loginTracking` is `.limited` are 'email', 'public_profile', 'gaming_profile' and 'gaming_user_picture'
+ @param tracking the tracking preference to use for a login attempt.
+ @param nonce an optional nonce to use for the login attempt. A valid nonce must be a non-empty string without whitespace.
+ Creation of the configuration will fail if the nonce is invalid. 
+ @param messengerPageId the associated page id  to use for a login attempt.
+ */
+- (nullable instancetype)initWithPermissions:(NSArray<NSString *> *)permissions
+                                    tracking:(FBSDKLoginTracking)tracking
+                                       nonce:(NSString *)nonce
+                             messengerPageId:(nullable NSString *)messengerPageId
+NS_REFINED_FOR_SWIFT;
 
 /**
  Attempts to initialize a new configuration with the expected parameters.
@@ -56,8 +75,21 @@ NS_SWIFT_NAME(LoginConfiguration)
  Creation of the configuration will fail if the nonce is invalid.
  */
 - (nullable instancetype)initWithPermissions:(NSArray<NSString *> *)permissions
-                         tracking:(FBSDKLoginTracking)tracking
+                                    tracking:(FBSDKLoginTracking)tracking
                                        nonce:(NSString *)nonce
+NS_REFINED_FOR_SWIFT;
+
+/**
+ Attempts to initialize a new configuration with the expected parameters.
+
+ @param permissions the requested permissions for the login attempt. Permissions must be an array of strings that do not contain whitespace.
+  The only permissions allowed when the `loginTracking` is `.limited` are 'email', 'public_profile', 'gaming_profile' and 'gaming_user_picture'
+ @param tracking the tracking preference to use for a login attempt.
+ @param messengerPageId the associated page id  to use for a login attempt.
+ */
+- (nullable instancetype)initWithPermissions:(NSArray<NSString *> *)permissions
+                                    tracking:(FBSDKLoginTracking)tracking
+                             messengerPageId:(nullable NSString *)messengerPageId
 NS_REFINED_FOR_SWIFT;
 
 /**
@@ -67,7 +99,7 @@ NS_REFINED_FOR_SWIFT;
  @param tracking the tracking preference to use for a login attempt.
  */
 - (nullable instancetype)initWithPermissions:(NSArray<NSString *> *)permissions
-                         tracking:(FBSDKLoginTracking)tracking
+                                    tracking:(FBSDKLoginTracking)tracking
 NS_REFINED_FOR_SWIFT;
 
 /**