+++ FBSDKLoginKit/FBSDKLoginKit/Swift/LoginConfiguration.swift
@@ -33,16 +33,19 @@ public extension LoginConfiguration {
    - parameter nonce: An optional nonce to use for the login attempt.
     A valid nonce must be an alphanumeric string without whitespace.
     Creation of the configuration will fail if the nonce is invalid. Defaults to a `UUID` string.
+   - parameter messengerPageId: An optional page id to use for a login attempt. Defaults to `nil`
    */
   convenience init?(
     permissions: Set<Permission> = [],
     tracking: LoginTracking = .enabled,
-    nonce: String = UUID().uuidString
+    nonce: String = UUID().uuidString,
+    messengerPageId: String? = nil
   ) {
     self.init(
       __permissions: permissions.map { $0.name },
       tracking: tracking,
-      nonce: nonce
+      nonce: nonce,
+      messengerPageId: messengerPageId
     )
   }
 }